services/user_service/pyproject.toml
[tool.poetry]
name = "user-service"
version = "0.1.0"
description = "User service for my-scaffold-project"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.60.0"
grpcio-tools = "^1.60.0"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.0"
uuid = "^1.30"
protobuf = "6.31.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
black = "^24.0.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 
